+++ src/main/java/in/koreatech/koin/domain/bus/util/BusCacheScheduler.java
@@ -3,26 +3,21 @@ package in.koreatech.koin.domain.bus.util;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-import in.koreatech.koin.domain.bus.util.city.CityBusClient;
-import in.koreatech.koin.domain.bus.util.city.CityBusRouteClient;
-import in.koreatech.koin.domain.bus.util.express.ExpressBusService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class BusScheduler {
+public class BusCacheScheduler {
 
-    private final ExpressBusService expressBusService;
-    private final CityBusClient cityBusClient;
-    private final CityBusRouteClient cityBusRouteClient;
+    private final BusCacheService busSchedulerService;
 
+    // TODO: 레거시 - 버스번호는 정적으로 유지함
     @Scheduled(cron = "0 * * * * *")
     public void cacheCityBusByOpenApi() {
         try {
-            cityBusClient.storeRemainTime();
-            cityBusRouteClient.storeCityBusRoute();
+            busSchedulerService.cacheCityBusByOpenApi();
         } catch (Exception e) {
             log.warn("시내버스 스케줄링 과정에서 오류가 발생했습니다.");
         }
@@ -31,7 +26,7 @@ public class BusScheduler {
     @Scheduled(cron = "0 30 0 * * *")
     public void cacheExpressBusByOpenApi() {
         try {
-            expressBusService.storeRemainTimeByRatio();
+            busSchedulerService.storeRemainTimeByRatio();
         } catch (Exception e) {
             log.warn("시외버스 스케줄링 과정에서 오류가 발생했습니다.");
         }
